Restructure src/data/object_events and src/data/field_effects
