+++ src/data/field_effects/field_effect_object_graphics.h
