+++ src/data/field_effects/field_effect_object_template_pointers.h
