+++ src/data/object_events/object_event_graphics_info_pointers.h
