+++ src/data/object_events/object_event_subsprites.h
